requirements.txt
requests==2.31.0
google-auth==2.25.1
google-api-python-client==2.115.0 